Keep only the factory class for simplicity
