--- csharp-selenium-webdriver-sample/BrowserType.cs
@@ -1,8 +1,0 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-public enum BrowserType
-{
-    Chrome,
-    Firefox,
-}
